move dsl datatype mirrors into datatypes/primitives/ (control.py renamed to vehicle.py)
